+++ src/compas/geometry/_core/tangent.py
@@ -22,13 +22,13 @@ def tangent_points_to_circle_xy(circle, point):
 
     Examples
     --------
-    >>> from compas.geometry import allclose
+    >>> from compas.tolerance import TOL
     >>> circle = ((0, 0, 0), (0, 0, 1)), 1.0
     >>> point = (2, 4, 0)
     >>> t1, t2 = tangent_points_to_circle_xy(circle, point)
-    >>> allclose(t1, [-0.772, 0.636, 0.000], 1e-3)
+    >>> TOL.is_allclose(t1, [-0.772, 0.636, 0.000], atol=1e-3)
     True
-    >>> allclose(t2, [0.972, -0.236, 0.000], 1e-3)
+    >>> TOL.is_allclose(t2, [0.972, -0.236, 0.000], atol=1e-3)
     True
 
     """