+++ src/compas/geometry/polygon.py
@@ -10,7 +10,6 @@ from compas.geometry import Line
 from compas.geometry import Plane
 from compas.geometry import Point
 from compas.geometry import Transformation
-from compas.geometry import allclose
 from compas.geometry import area_polygon
 from compas.geometry import bounding_box
 from compas.geometry import centroid_polygon
@@ -19,6 +18,7 @@ from compas.geometry import is_coplanar
 from compas.geometry import is_polygon_convex
 from compas.geometry import transform_points
 from compas.itertools import pairwise
+from compas.tolerance import TOL
 
 
 class Polygon(Geometry):
@@ -119,7 +119,7 @@ class Polygon(Geometry):
     def __eq__(self, other):
         if not hasattr(other, "__iter__") or not hasattr(other, "__len__") or len(self) != len(other):
             return False
-        return allclose(self, other)
+        return TOL.is_allclose(self, other)
 
     # ==========================================================================
     # Properties